PMI-587 Rename venv to indicate which repo it belongs to
The PA's have multiple terminal windows open at a time when manually
running reports. The default bash profile on OSX does not indicate the
full folder path so at present they have no way to tell these windows
apart and which virtualenv each window is running in.

The virtualenv name always appears at the start of the terminal input
line, so renaming the venv to include the repo name ensures they can
tell which venv is which.