PMI-648: Use stub `Config` when running unit tests.
Setting the environment variable `ENV` as 'test' will signal that tests
are being run and `TestConfig` class should be used in lieu of `Config`.

The code has also been refactored to use `Config` where
`RPFederationConfig` was being used. This is justified by the following
reasons:

1. `RPFederationConfig` only provides the RP mapping, which `Config`
already has (this refactor was also marked as a TODO during previous
quarter).

2. This simplifies unit tests as now only `Config` needs to be tested
and can easily be mocked/stubbed/faked/etc.

3. Keeping `RPFederationConfig` would have required changes to its test,
which would eventually had to be removed anyway if we refactored it at a
later stage. So, the test has been updated to cover `Config` instead.