PMI-753: Read Google auth credentials from environment variables.
The implementation is quite similar to how we do this for Piwik token.
Since `Pygsheets` doesn't currently support reading credentials from an
in-memory data structure, we have had to implement creating a temporary
credentials file to do the job. We may revisit this later after the
next major release of `Pygsheets`.

Also note that, at present, `dotenv` package doesn't support multi-line
environment variables properly. Therefore, when running locally, the
environment variable `GOOGLE_AUTH_PRIVATE_KEY` cannot be defined in
`.env` file and must be defined in the shell. There is a PR currently
open on Github repo of `dotenv` which fixes this behaviour, so we'll
revisit this later.